+++ src/symfc/basis_sets/basis_sets_O4.py
@@ -251,7 +251,7 @@ class FCBasisSetO4(FCBasisSetBase):
         if self._fc_cutoff is None:
             n_sym, N = self._spg_reps._permutations.shape
             basis_size_estimates = 81 * (N**4) / n_sym / 24
-            return np.round(basis_size_estimates).astype(int)
+            return int(np.round(basis_size_estimates).astype(int))
 
         trans_perms = self._spg_reps.translation_permutations
         c_pt = compr_permutation_lat_trans_O4(
@@ -262,4 +262,4 @@ class FCBasisSetO4(FCBasisSetBase):
         )
         n_sym_prim = len(self._spg_reps._unique_rotations)
         basis_size_estimates = c_pt.shape[1] / n_sym_prim
-        return np.round(basis_size_estimates).astype(int)
+        return int(np.round(basis_size_estimates).astype(int))
